+++ src/containers/Template/Form/Template.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import * as Yup from 'yup';
 import { Input } from '../../../components/UI/Form/Input/Input';
 import { EmojiInput } from '../../../components/UI/Form/EmojiInput/EmojiInput';
-import { FormItem } from '../../Form/Form';
+import { FormItem } from '../../Form/FormLayout';
 import { GET_TEMPLATE } from '../../../graphql/queries/Template';
 import {
   CREATE_TEMPLATE,
